added role based access control to all routes

+++ api/routes/tournaments.js
@@ -1,7 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const connection = require('../connection');
+const mysqlConnection = require("../../connection");
 const Tournament = require('../models/tournaments');
+const authenticate = require("../middleware/authenticate");
+const authorize = require("../middleware/authorize");
+const Role = require("../models/role");
 
 // Get all tournaments
 router.get('/', (req, res) => {
@@ -21,35 +24,40 @@ router.get('/', (req, res) => {
     })
 });
 
-router.post("/", (req, res, next) => {
-  const newTournament = new Tournament({
-    id: null,
-    name: req.body.name,
-    description: req.body.description,
-    startTime: req.body.startTime,
-    endTime: req.body.endTime,
-    roundRobinDuration: req.body.roundRobinDuration,
-    playoffDuration: req.body.playoffDuration,
-    quarterFinalDuration: req.body.quarterFinalDuration,
-    semiFinalDuration: req.body.semiFinalDuration,
-    finalDuration: req.body.finalDuration,
-  });
-  const query = "INSERT INTO tournaments SET ?";
+router.post(
+  "/",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const newTournament = new Tournament({
+      id: null,
+      name: req.body.name,
+      description: req.body.description,
+      startTime: req.body.startTime,
+      endTime: req.body.endTime,
+      roundRobinDuration: req.body.roundRobinDuration,
+      playoffDuration: req.body.playoffDuration,
+      quarterFinalDuration: req.body.quarterFinalDuration,
+      semiFinalDuration: req.body.semiFinalDuration,
+      finalDuration: req.body.finalDuration,
+    });
+    const query = "INSERT INTO tournaments SET ?";
 
-  connection.query(query, newTournament, (err, result) => {
-    if (err) {
-      console.log("Inserting new tournament failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully inserted new tournament.");
-      res.status(201).json({
-        message: "Successfully inserted new tournament.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, newTournament, (err, result) => {
+      if (err) {
+        console.log("Inserting new tournament failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully inserted new tournament.");
+        res.status(201).json({
+          message: "Successfully inserted new tournament.",
+        });
+      }
+    });
+  }
+);
   
 router.get('/:tournamentId', (req, res, next) => {
   const query = "SELECT * FROM tournaments WHERE id = " + req.params.tournamentId;
@@ -100,41 +108,67 @@ router.get("/:tournamentId/teams", (req, res, next) => {
   });
 });
 
-  
-router.patch('/:tournamentId', (req, res, next) => {
-  const query =
-  "UPDATE tournaments SET ? where id = " + req.params.tournamentId;
-
-  connection.query(query, req.body, (err, result) => {
+router.get("/:tournamentId/players", (req, res, next) => {
+  const query = "SELECT * FROM divisions WHERE tournamentId = " + req.params.tournamentId;
+  mysqlConnection.query(query, (err, rows, fields) => {
     if (err) {
-      console.log("Patching tournament failed.");
+      console.log("Getting divisions from tournament failed.");
       res.status(500).json({
         error: err,
       });
     } else {
-      console.log("Successfully patched tournament.");
-      res.status(200).json({
-        message: "Successfully patched tournament.",
-      });
-      }
-    });
-  });
-  
-router.delete('/:tournamentId', (req, res, next) => {
-  const query = "DELETE FROM tournaments WHERE id = " + req.params.tournamentId;
-  connection.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log("Deleting tournament failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully deleted tournament.");
-      res.status(200).json({
-        message: "Successfully deleted tournament."
-      });
+      console.log("Successfully received divisions from tournament.");
+      res.status(200).json(rows);
     }
   });
 });
+  
+router.patch(
+  "/:tournamentId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query =
+      "UPDATE tournaments SET ? where id = " + req.params.tournamentId;
+
+      mysqlConnection.query(query, req.body, (err, result) => {
+        if (err) {
+          console.log("Patching tournament failed.");
+          res.status(500).json({
+            error: err,
+          });
+        } else {
+          console.log("Successfully patched tournament.");
+          res.status(200).json({
+            message: "Successfully patched tournament.",
+          });
+        }
+      });
+    }
+  );
+  
+  router.delete(
+    "/:tournamentId",
+    authenticate,
+    authorize([Role.Admin, Role.Owner]),
+    (req, res, next) => {
+      const query =
+        "DELETE FROM tournaments WHERE id = " + req.params.tournamentId;
+  
+      mysqlConnection.query(query, (err, rows, fields) => {
+        if (err) {
+          console.log("Deleting tournament failed.");
+          res.status(500).json({
+            error: err,
+          });
+        } else {
+          console.log("Successfully deleted tournament.");
+          res.status(200).json({
+            message: "Successfully deleted tournament.",
+          });
+        }
+      });
+    }
+  );
 
 module.exports = router;
